restructure to avoid running code multiple times
