+++ scripts/examples.py
@@ -1,12 +1,8 @@
-import os
-import shutil
-from os import listdir
 from pathlib import Path
 
-from generate import generate_db
-from shared import EXAMPLES_PATH, get_package
+from shared import gmns
 
-package = get_package()
+EXAMPLES_PATH = gmns.EXAMPLES_PATH
 
 # List of tuples with the database name and path to the relevant example folder
 EXAMPLE_PATHS: list[tuple[str, Path]] = [
@@ -20,30 +16,6 @@ FILES_TO_COPY: list[tuple[str, Path]] = [
 ]
 
 
-def prep_examples(
-    examples: list[tuple[str, Path]] | None = None,
-    copy_files: list[tuple[str, Path]] | None = None,
-):
-    if examples is None:
-        examples = []
-
-    if copy_files is None:
-        copy_files = []
-
-    files_to_delete: list[Path] = []
-    for example_name, example_path in examples:
-        for name_of_file_to_copy, path_of_file_to_copy in copy_files:
-            if name_of_file_to_copy not in listdir(example_path):
-                shutil.copyfile(
-                    path_of_file_to_copy, example_path / name_of_file_to_copy
-                )
-                files_to_delete.append(example_path / name_of_file_to_copy)
-        generate_db(example_path, example_name, False)
-
-    for file_to_delete in files_to_delete:
-        os.remove(file_to_delete)
-
-
 if __name__ == "__main__":
-    prep_examples(EXAMPLE_PATHS, FILES_TO_COPY)
+    gmns.prep_examples(EXAMPLE_PATHS, FILES_TO_COPY)
     pass
